pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "python-dotenv>=1.1.1",
]
